refactor: consistent mention event hook with other event hooks
